Refactored ModelEditorToolbar into its own file
