+++ CMakeLists.txt
@@ -506,6 +506,8 @@ add_library(osccore STATIC
     src/OpenSimBindings/Widgets/ModelActionsMenuItems.hpp
     src/OpenSimBindings/Widgets/ModelEditorMainMenu.cpp
     src/OpenSimBindings/Widgets/ModelEditorMainMenu.hpp
+    src/OpenSimBindings/Widgets/ModelEditorToolbar.cpp
+    src/OpenSimBindings/Widgets/ModelEditorToolbar.hpp
     src/OpenSimBindings/Widgets/ModelMusclePlotPanel.cpp
     src/OpenSimBindings/Widgets/ModelMusclePlotPanel.hpp
     src/OpenSimBindings/Widgets/NavigatorPanel.cpp
