+++ src/OpenSimBindings/Widgets/ModelEditorToolbar.hpp
@@ -1,0 +1,29 @@
+#pragma once
+
+#include <memory>
+#include <string_view>
+
+namespace osc { class MainUIStateAPI; }
+namespace osc { class UndoableModelStatePair; }
+
+namespace osc
+{
+    class ModelEditorToolbar final {
+    public:
+        ModelEditorToolbar(
+            std::string_view,
+            MainUIStateAPI*,
+            std::shared_ptr<UndoableModelStatePair>);
+        ModelEditorToolbar(ModelEditorToolbar const&) = delete;
+        ModelEditorToolbar(ModelEditorToolbar&&) noexcept;
+        ModelEditorToolbar& operator=(ModelEditorToolbar const&) = delete;
+        ModelEditorToolbar& operator=(ModelEditorToolbar&&) noexcept;
+        ~ModelEditorToolbar() noexcept;
+
+        void draw();
+
+    private:
+        class Impl;
+        std::unique_ptr<Impl> m_Impl;
+    };
+}
